refactor: Extract draft features.
Draft features are account level
